Allow `mullvad-problem-report` to be used as a lib
